Update main.py
Бады:
1. Исправил в строке 186 название переменной - неверно указал
2. Сделал проверку на существование файла: если он есть то обновляеться, если нет создаеться
3. Вставил комменты ко всем функциям

+++ main.py
@@ -2,11 +2,12 @@
 import pandas as pd
 import openpyxl
 import datetime
+import os
 
 def import_to_main_base_csv(path_file):
     """функция импортирует из файла csv данные и записывает из в excel файл"""
     data_csv = pd.read_csv(path_file)
-    writer = pd.ExcelWriter('base_main.xlsx', mode="a", engine='openpyxl',if_sheet_exists='overlay')
+    writer = pd.ExcelWriter('base_main.xlsx', mode="a", engine='openpyxl', if_sheet_exists='overlay')
     data_csv.to_excel(writer, 'marks')
     writer.save()
     writer.close()
@@ -15,7 +16,7 @@ def import_to_main_base_csv(path_file):
 def import_to_main_base_json(path_file):
     """функция импортирует из файла json данные и записывает из в excel файл"""
     data_json = pd.read_json(path_file)
-    writer = pd.ExcelWriter('base_main.xlsx',mode="a", engine='openpyxl',if_sheet_exists='overlay')
+    writer = pd.ExcelWriter('base_main.xlsx', mode="a", engine='openpyxl', if_sheet_exists='overlay')
     data_json.to_excel(writer, 'marks')
     writer.save()
     writer.close()
@@ -24,13 +25,14 @@ def import_to_main_base_json(path_file):
 def import_to_main_base_excel(path_file):
     """функция импортирует из файла excel данные и записывает из в excel файл"""
     data_json = pd.read_excel(path_file)
-    writer = pd.ExcelWriter('base_main.xlsx', mode="a", engine='openpyxl',if_sheet_exists='overlay')
+    writer = pd.ExcelWriter('base_main.xlsx', mode="a", engine='openpyxl', if_sheet_exists='overlay')
     data_json.to_excel(writer, 'marks')
     writer.save()
     writer.close()
 
 
 def enter_data_from_keyboard(list):
+    """Функция ввода данных вручную"""
     df = pd.DataFrame(list, index=[0])
     df.head()
     excel_data_df = pd.read_excel('base_main.xlsx', index_col=0, sheet_name='marks')
@@ -41,7 +43,9 @@ def enter_data_from_keyboard(list):
     writer.save()
     writer.close()
 
+
 def caunter_age(year_naw):
+    """Функция формирования отчета возраст"""
     excel_data_df = pd.read_excel('base_main.xlsx', index_col=0, sheet_name='marks')
     excel_data_df.insert(10, "Year_Calc", year_naw)
     excel_data_df["Age"] = (excel_data_df["Year_Calc"] - excel_data_df["birt_year"])
@@ -50,21 +54,27 @@ def caunter_age(year_naw):
     df_unit = pd.concat([excel_data_df_live, excel_data_df_death], ignore_index=True)
     return df_unit
 
+
 def caunter_age_live(year_naw):
+    """Функция формирования отчета вывода живых"""
     excel_data_df = pd.read_excel('base_main.xlsx', index_col=0, sheet_name='marks')
     excel_data_df.insert(10, "Year_Calc", year_naw)
     excel_data_df["Age"] = (excel_data_df["Year_Calc"] - excel_data_df["birt_year"])
     excel_data_df_live = excel_data_df.loc[excel_data_df['live'] == "Да"]
     return excel_data_df_live
 
+
 def caunter_age_death(year_naw):
+    """Функция формирования отчета вывода мертвых"""
     excel_data_df = pd.read_excel('base_main.xlsx', index_col=0, sheet_name='marks')
     excel_data_df.insert(10, "Year_Calc", year_naw)
     excel_data_df["Age"] = (excel_data_df["Year_Calc"] - excel_data_df["birt_year"])
     excel_data_df_death = excel_data_df.loc[excel_data_df['live'] == "Нет"]
     return excel_data_df_death
 
+
 def seacher_text(txt):
+    """Функция поиска по тексту """
     excel_data_df = pd.read_excel('base_main.xlsx', index_col=0, sheet_name='marks')
     search_surname_df = excel_data_df[excel_data_df['surname'].str.contains(txt)]
     search_name_df = excel_data_df[excel_data_df['name'].str.contains(txt)]
@@ -72,14 +82,25 @@ def seacher_text(txt):
     df_unit1 = pd.concat([search_surname_df, search_name_df], ignore_index=True)
     df_unit2 = (pd.concat([df_unit1, search_patronymic_df], ignore_index=True)).drop_duplicates()
     return df_unit2
-df = pd.DataFrame(columns=(
-['surname', 'name', 'patronymic','birt_year', 'birth_month', 'birth_day', 'live', 'death_year', 'death_month',
- 'death_day','gender']))
-writer = pd.ExcelWriter('base_main.xlsx', mode="a", engine='openpyxl',if_sheet_exists='overlay')
-df.to_excel(writer,'marks')
-writer.save()
-writer.close()
-
+"""Старт программы"""
+"""Блок проверки на существование файла"""
+if os.path.exists("/home/volodymyr/pythonProject/pythonProject123/base_main.xlsx"):
+    df = pd.DataFrame(columns=(
+        ['surname', 'name', 'patronymic', 'birt_year', 'birth_month', 'birth_day', 'live', 'death_year', 'death_month',
+         'death_day', 'gender']))
+    writer = pd.ExcelWriter('base_main.xlsx', mode="a", engine='openpyxl', if_sheet_exists='overlay')
+    df.to_excel(writer, 'marks')
+    writer.save()
+    writer.close()
+else:
+    df = pd.DataFrame(columns=(
+        ['surname', 'name', 'patronymic', 'birt_year', 'birth_month', 'birth_day', 'live', 'death_year',
+         'death_month',
+         'death_day', 'gender']))
+    writer = pd.ExcelWriter('base_main.xlsx', mode="w", engine='openpyxl')
+    df.to_excel(writer, 'marks')
+    writer.save()
+    writer.close()
 while True:
     base_ask = int(input(
         "Вас приветствует программа обработки персональных данных\n Для ввода данных нажмите 1 \n Для обработки персональных данных нажмите 2\n"))
@@ -105,7 +126,8 @@ while True:
                 continue
         if choice == 2:
             choice_var = int(
-                input("выбиретет вариант ввода данных \n 1- ввод данных в одну строку \n 2-ввод данных последовательно\n"))
+                input(
+                    "выбиретет вариант ввода данных \n 1- ввод данных в одну строку \n 2-ввод данных последовательно\n"))
             if choice_var == 2:
                 live_inp = input("Человек на данный момент жив? \n Веедите Да или Нет\n")
                 test_isdigit = live_inp.isdigit()
@@ -115,11 +137,11 @@ while True:
                         len_inp_death_year = len(death_year_inp)
                         test_isdigit_death_year = death_year_inp.isdigit()
                         if not test_isdigit_death_year:
-                            print("вы ввели не число, повторите ввод")
+                            print("вы ввели не число, повторите ввод\n")
                             continue
                         death_year_inp = int(death_year_inp)
                         if len_inp_death_year > 4:
-                            print(("вы ввели некорректный год, цифр должно быть 4"))
+                            print(("вы ввели некорректный год, цифр должно быть 4\n"))
                             continue
                         else:
                             break
@@ -132,7 +154,7 @@ while True:
                             continue
                         death_month_inp = int(death_month_inp)
                         if death_month_inp > 12:
-                            print(("вы ввели некорректный месяц, в году месяцев 12"))
+                            print(("вы ввели некорректный месяц, в году месяцев 12\n"))
                             continue
                         else:
                             break
@@ -141,24 +163,24 @@ while True:
                         len_death_day_inp = len(death_day_inp)
                         test_isdigit_death_day_inp = death_day_inp.isdigit()
                         if not test_isdigit_death_day_inp:
-                            print("вы ввели не число, повторите ввод")
+                            print("вы ввели не число, повторите ввод\n")
                             continue
                         death_day_inp_int = int(death_day_inp)
                         if death_day_inp_int > 31:
-                            print(("вы ввели некорректный день, в месяце максимум 31 день"))
+                            print(("вы ввели некорректный день, в месяце максимум 31 день\n"))
                             continue
                         else:
                             break
                 surname_inp = input(("Введите фамилию \n").title())
                 name_inp = input(("Введите имя \n").title())
                 patronymic_inp = input(("Введите отчество \n").title())
-                gender_inp = input("Введите пол: М- мужской, Ж-женский")
+                gender_inp = input("Введите пол: М- мужской, Ж-женский\n")
                 while True:
                     birt_year_inp = input("Введите год рождения \n")
                     len_birt_year_inp = len(birt_year_inp)
                     test_birt_year_inp = birt_year_inp.isdigit()
                     if not test_birt_year_inp:
-                        print("вы ввели не число, повторите ввод")
+                        print("вы ввели не число, повторите ввод\n")
                         continue
                     birt_year_inp = int(birt_year_inp)
                     if len_birt_year_inp > 4:
@@ -180,7 +202,7 @@ while True:
                         break
                 while True:
                     birth_day_inp = input("Введите день рождения \n")
-                    test_isdigit_birth_day_inp = death_day_inp.isdigit()
+                    test_isdigit_birth_day_inp = birth_day_inp.isdigit()
                     if not test_isdigit_birth_day_inp:
                         print("вы ввели не число, повторите ввод")
                         continue
@@ -197,7 +219,8 @@ while True:
             if choice_var == 1:
                 live_inp = input("Человек на данный момент жив? \n Веедите Да или Нет\n")
                 if live_inp == "Да":
-                    var_birthday_input = input('Выберете даты рождения  в следующем формате: \n  дд.мм.год  или \n дд мм год или \n дд/мм/год\n дд-мм-год')
+                    var_birthday_input = input(
+                        'Выберете даты рождения  в следующем формате: \n  дд.мм.год  или \n дд мм год или \n дд/мм/год\n дд-мм-год')
                     birth_day_inp = int(var_birthday_input[0:2])
                     birth_month_inp = int(var_birthday_input[3:5])
                     birt_year_inp = int(var_birthday_input[6:10])
@@ -209,11 +232,13 @@ while True:
                     patronymic_inp = input(("Введите отчество \n").title())
                     gender_inp = input("Введите пол: М- мужской, Ж-женский\n")
                 if live_inp == "Нет":
-                    var_birthday_input = input("Выберете даты рождения  в следующем формате: \n  дд.мм.год  или \n дд мм год или \n дд/мм/год\n дд-мм-год\n")
+                    var_birthday_input = input(
+                        "Выберете даты рождения  в следующем формате: \n  дд.мм.год  или \n дд мм год или \n дд/мм/год\n дд-мм-год\n")
                     birth_day_inp = int(var_birthday_input[0:2])
                     birth_month_inp = int(var_birthday_input[3:5])
                     birt_year_inp = int(var_birthday_input[6:10])
-                    var_death_input1 = input("Выберете дату смерти  в следующем формате: \n  дд.мм.год  или \n дд мм год или \n дд/мм/год\n дд-мм-год\n")
+                    var_death_input1 = input(
+                        "Выберете дату смерти  в следующем формате: \n  дд.мм.год  или \n дд мм год или \n дд/мм/год\n дд-мм-год\n")
                     death_year_inp = int(var_death_input1[0:2])
                     death_month_inp = int(var_death_input1[3:5])
                     death_day_inp = int(var_death_input1[6:10])
@@ -223,12 +248,14 @@ while True:
                     gender_inp = input("Введите пол: М- мужской, Ж-женский\n")
             else:
                 break
-        n = ({"surname": surname_inp, "name": name_inp, "patronymic": patronymic_inp,  "birt_year": birt_year_inp,
-              "birth_month": birth_month_inp, "birth_day": birth_day_inp, "live": live_inp, "death_year": death_year_inp,
+        n = ({"surname": surname_inp, "name": name_inp, "patronymic": patronymic_inp, "birt_year": birt_year_inp,
+              "birth_month": birth_month_inp, "birth_day": birth_day_inp, "live": live_inp,
+              "death_year": death_year_inp,
               "death_month": death_month_inp, "death_day": death_day_inp, "gender": gender_inp})
         enter_data_from_keyboard(n)
     if base_ask == 2:
-        choice_rep = int(input("Приветствую Вас в системе обработки данных и их вывода \n 1- вывести возраст людей внесенных в базу\n 2- вывести данне живущих людей \n 3 -вывести данные мертвых \n 4- поиск и вывод данных по всей базе \n"))
+        choice_rep = int(input(
+            "Приветствую Вас в системе обработки данных и их вывода \n 1- вывести возраст людей внесенных в базу\n 2- вывести данне живущих людей \n 3 -вывести данные мертвых \n 4- поиск и вывод данных по всей базе \n"))
         if choice_rep == 1:
             yaer_rep = int(input("Введите год на который нужно узнать возраст всех людей в базе\n"))
             rep1 = caunter_age(yaer_rep)
@@ -250,5 +277,3 @@ while True:
         continue
     if ask == "Нет":
         break
-
-
